models/operations→app/usecasesディレクトリに移動
