+++ app/graphql/admin/v1/mutations/create_order.rb
@@ -11,7 +11,7 @@ module Admin
         field :order, Types::OrderType, null: false
 
         def resolve(**args)
-          order = Operations::CreateOrder.execute(args)
+          order = CreateOrderAndAdjustProductStock.execute(args)
           raise GraphQL::ExecutionError, '注文を受け付けることが出来ません' unless order
 
           { order: order }
