+++ spec/usecases/create_order_and_adjust_product_stock_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe Operations::CreateOrder, type: :model do
+RSpec.describe CreateOrderAndAdjustProductStock, type: :model do
   describe '.execute' do
     let(:customer) { create(:customer) }
     let(:product) { create(:product, price: 1000) }
